Move `check.capsule` sub-pkg to `check.template`.
