+++ at-paninij-core/at-paninij-proc/src/main/java/org/paninij/proc/check/template/SuffixCheck.java
@@ -1,4 +1,4 @@
-package org.paninij.proc.check.capsules;
+package org.paninij.proc.check.template;
 
 import javax.lang.model.element.TypeElement;
